+++ apps/api/src/schemas/review.py
@@ -4,7 +4,7 @@ from typing import List
 
 class ReviewCreate(BaseModel):
     washroom_id: str
-    rating: int = Field(ge=1, le=5)
+    rating: float = Field(ge=1, le=5)
     comment: str | None = None
 
 
@@ -12,5 +12,5 @@ class ReviewOut(BaseModel):
     id: str
     washroom_id: str
     user_id: str
-    rating: int
+    rating: float
     comment: str | None
